Fix tests.
We need to get the actual nodeId that has been assigned by the monitor,
rather than our own counter: in case of failure to create a node we might
consume some sequence numbers that we are never going to get back.

To that end fix the command `pg_autoctl do fsm state` for it to only print
the current state, rather than trying to be smart and update what can be.
Now that Postgres only runs when pg_autoctl runs, wanting to have a running
pg_autoctl to be able to debug it is kind of lame.